Make protocol hash size configurable
